move user management into user_system extension
This allows developers to do
  config.extensions = [:my_replacement_user_system, :all]
  canfig.extensions -= [:user_system]
which will allow a completely custom login and user
management solution. Currently, the bootstrap process
still relies on the existance of a User model. This
should be changed to allow some altered bootstrap process
such as loading seed data or templates from extensions.

An additional caveat is that the migrations are still in
the main project's migrations rather than the extension's.
A separate bit of code will be required to address this.